+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - [TD-4100] Allow partial loads for notes in case that there are errors on the file
+- [TD-4293] Added quality implementations audit events
 
 ## [4.35.1] 2022-01-10